requirements.txt
fastapi
uvicorn[standard]
python-multipart
SQLAlchemy
passlib[bcrypt]
python-jose[cryptography]
alembic
psycopg2-binary

selenium
google-generativeai
apify-client